# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KeywordCreate": {
        "properties": {
          "keyword": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Keyword",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 280,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "severity": {
            "enum": [
              "HARMFUL",
              "OFFENSIVE"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "keyword",
          "severity"
        ],
        "title": "KeywordCreate",
        "type": "object"
      },
      "KeywordEntry": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "keyword": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Keyword",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 280,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "severity": {
            "enum": [
              "HARMFUL",
              "OFFENSIVE"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "keyword",
          "severity"
        ],
        "title": "KeywordEntry",
        "type": "object"
      },
...
      "RenameSelf": {
        "properties": {
          "display_name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          }
        },
        "required": [
          "display_name"
        ],
        "title": "RenameSelf",
        "type": "object"
      },
      "RenameWorkspace": {
        "properties": {
          "name": {
            "maxLength": 80,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "RenameWorkspace",
        "type": "object"
      },
...
      "SetAutoDeleteDelay": {
        "properties": {
          "delay_minutes": {
            "enum": [
              0,
              5,
              30,
              60,
              720,
              1440
            ],
            "title": "Delay Minutes",
            "type": "integer"
          }
        },
        "required": [
          "delay_minutes"
        ],
        "title": "SetAutoDeleteDelay",
        "type": "object"
      },
      "SetKeywordList": {
        "properties": {
          "keywords": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "title": "Keywords",
            "type": "array"
          }
        },
        "required": [
          "keywords"
        ],
        "title": "SetKeywordList",
        "type": "object"
      },
      "SetToggle": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "SetToggle",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkspaceSettings": {
        "properties": {
          "auto_delete_delay_minutes": {
            "title": "Auto Delete Delay Minutes",
            "type": "integer"
          },
          "auto_hide_offensive": {
            "title": "Auto Hide Offensive",
            "type": "boolean"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "is_sandbox": {
            "title": "Is Sandbox",
            "type": "boolean"
          },
          "keyword_allowlist": {
            "items": {
              "type": "string"
            },
            "title": "Keyword Allowlist",
            "type": "array"
          },
          "keyword_blocklist": {
            "items": {
              "type": "string"
            },
            "title": "Keyword Blocklist",
            "type": "array"
          },
          "sample_comments": {
            "title": "Sample Comments",
            "type": "integer"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          },
          "workspace_name": {
            "title": "Workspace Name",
            "type": "string"
          },
          "your_role": {
            "title": "Your Role",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "workspace_name",
          "is_sandbox",
          "your_role",
          "display_name",
          "sample_comments",
          "auto_delete_delay_minutes",
          "auto_hide_offensive",
          "keyword_allowlist",
          "keyword_blocklist"
        ],
        "title": "WorkspaceSettings",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "KCMS Backend",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/settings": {
      "get": {
        "operationId": "getSettings",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Settings"
      }
    },
    "/api/v1/settings/auto-delete": {
      "patch": {
        "description": "How long a HARMFUL comment is quarantined before KCMS deletes it from\nFacebook. Governs the whole workspace's Pages, so only an owner sets it —\nsame rule as renaming the workspace itself.",
        "operationId": "setAutoDeleteDelay",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAutoDeleteDelay"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Auto Delete Delay"
      }
    },
    "/api/v1/settings/auto-hide-offensive": {
      "patch": {
        "description": "Whether an OFFENSIVE comment is hidden on Facebook immediately, ahead\nof any human review. Governs the whole workspace, so owner-only.",
        "operationId": "setAutoHideOffensive",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetToggle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Auto Hide Offensive"
      }
    },
    "/api/v1/settings/keyword-allowlist": {
      "patch": {
        "description": "Phrases that force SAFE, ahead of the blocklist and the pattern\nmatcher's own vocabulary alike.",
        "operationId": "setKeywordAllowlist",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetKeywordList"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Keyword Allowlist"
      }
    },
    "/api/v1/settings/keyword-blocklist": {
      "patch": {
        "description": "Phrases that force HARMFUL. Loses to an allowlist match on the same\ncomment, but otherwise outranks the pattern matcher.",
        "operationId": "setKeywordBlocklist",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetKeywordList"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Keyword Blocklist"
      }
    },
    "/api/v1/settings/keywords": {
      "get": {
        "description": "This workspace's own keywords. Never another workspace's, and never the\nshipped defaults: those are not editable and not listed as if they were.",
        "operationId": "getKeywords",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KeywordEntry"
                  },
                  "title": "Response Getkeywords",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Keywords"
      },
      "post": {
        "operationId": "addKeyword",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeywordCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeywordEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Keyword"
      }
    },
    "/api/v1/settings/keywords/{keyword}": {
      "delete": {
        "operationId": "removeKeyword",
        "parameters": [
          {
            "in": "path",
            "name": "keyword",
            "required": true,
            "schema": {
              "title": "Keyword",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Keyword"
      }
    },
    "/api/v1/settings/me": {
      "patch": {
        "description": "Anyone may change their own display name. It affects how future actions\nare attributed; past actions keep the name used at the time.",
        "operationId": "renameSelf",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameSelf"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Self"
      }
    },
    "/api/v1/settings/workspace": {
      "patch": {
        "operationId": "renameWorkspace",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameWorkspace"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Workspace"
      }
    },
...
  }
}